requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
yara-python==4.3.1
python-telegram-bot==20.7
pydantic==2.5.0
aiohttp==3.9.1
python-dotenv==1.0.0